refactor: move class
* Creating a new passcode package to later add a PasscodeEndpoint class to.
