+++ server/src/main/java/org/cloudfoundry/identity/uaa/passcode/PasscodeInformation.java
@@ -10,7 +10,7 @@
  *     subcomponents is subject to the terms and conditions of the
  *     subcomponent's license, as noted in the LICENSE file.
  *******************************************************************************/
-package org.cloudfoundry.identity.uaa.login;
+package org.cloudfoundry.identity.uaa.passcode;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
